requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-phonenumber-field==2.4.0
phonenumbers==8.10.11
djangorestframework==3.9.3
markdown==3.1
django-filter==2.1.0
django-multiselectfield==0.1.8
django-cors-headers==2.5.3
